refactor(embed): extract index/ai/providers/ — single resolution chokepoint (#317)
Task 1 of 7 toward the Ollama embedding backend (#317). PURE mechanical move, zero
behavior change — byte-green: 930 -> 930 tests. Centralizes all embedding-provider code
under one directory so "no provider code outside providers/" is true by construction, not
discipline.

- providers/mod.rs: the Embedder trait + embedder_for_spec dispatch + active_embedder
  resolution (signatures UNCHANGED) + the feature-missing consts + MockEmbedder.
- providers/{fastembed,model2vec,hash}.rs: the existing impls, one backend per file.
- ai/mod.rs curates the surface via `pub use providers::{...}` (mod.rs as the index, per
  rust-modern-style) — no transition shim.
- helpers.rs keeps embed_query/embed_query_with (now resolving active_embedder from its
  new home).

Verified: nextest 930 default / 934 --features eval; clippy --all-targets clean (default
AND --no-default-features, -D warnings); cargo +nightly fmt; workspace build.
